+++ include/hsmcpp/HsmEventDispatcherBase.hpp
@@ -5,6 +5,8 @@
 #define __HSMCPP_HSMEVENTDISPATCHERBASE_HPP__
 
 #include <map>
+#include <list>
+#include <mutex>
 #include "IHsmEventDispatcher.hpp"
 
 namespace hsmcpp
@@ -23,6 +25,10 @@ protected:
 public:
     virtual ~HsmEventDispatcherBase();
 
+    virtual HandlerID_t registerEventHandler(const EventHandlerFunc_t& handler) override;
+    virtual void unregisterEventHandler(const HandlerID_t handlerID) override;
+    virtual void emitEvent(const HandlerID_t handlerID) = 0;
+
     virtual HandlerID_t registerTimerHandler(const TimerHandlerFunc_t& handler) override;
     virtual void unregisterTimerHandler(const HandlerID_t handlerID) override;
 
@@ -37,16 +43,25 @@ public:
 protected:
     virtual HandlerID_t getNextHandlerID();
 
+    void unregisterAllEventHandlers();
+
     TimerInfo getTimerInfo(const TimerID_t timerID) const;
     TimerHandlerFunc_t getTimerHandlerFunc(const HandlerID_t handlerID) const;
 
     virtual void startTimerImpl(const TimerID_t timerID, const unsigned int intervalMs, const bool isSingleShot);
     virtual void stopTimerImpl(const TimerID_t timerID);
 
-private:
+    void dispatchPendingEvents();
+
+protected:
     HandlerID_t mNextHandlerId = 1;
     std::map<TimerID_t, TimerInfo> mActiveTimers;
+    std::map<HandlerID_t, EventHandlerFunc_t> mEventHandlers;
     std::map<HandlerID_t, TimerHandlerFunc_t> mTimerHandlers;
+    std::list<HandlerID_t> mPendingEvents;
+    std::mutex mEmitSync;
+    std::mutex mHandlersSync;
+
 };
 
 }// namespace hsmcpp