+++ include/hsmcpp/HsmEventDispatcherQt.hpp
@@ -5,10 +5,6 @@
 #define __HSMCPP_HSMEVENTDISPATCHERQT_HPP__
 
 #include "HsmEventDispatcherBase.hpp"
-#include <map>
-#include <mutex>
-#include <functional>
-
 #include <QObject>
 #include <QEvent>
 
@@ -26,19 +22,13 @@ public:
 
     bool start() override;
 
-    HandlerID_t registerEventHandler(const EventHandlerFunc_t& handler) override;
-    void unregisterEventHandler(const HandlerID_t handlerID) override;
-    void emitEvent() override;
+    void emitEvent(const HandlerID_t handlerID) override;
 
 protected:
-    void unregisterAllEventHandlers();
     bool event(QEvent* ev) override;
 
 private:
     static QEvent::Type mQtEventType;
-
-    std::mutex mHandlersSync;
-    std::map<HandlerID_t, EventHandlerFunc_t> mEventHandlers;
 };
 
 } // namespace hsmcpp