+++ include/hsmcpp/HsmEventDispatcherSTD.hpp
@@ -5,11 +5,7 @@
 #define __HSMCPP_HSMEVENTDISPATCHERSTD_HPP__
 
 #include "HsmEventDispatcherBase.hpp"
-#include <map>
-#include <mutex>
 #include <thread>
-#include <memory>
-#include <functional>
 #include <condition_variable>
 
 namespace hsmcpp
@@ -21,9 +17,7 @@ public:
     HsmEventDispatcherSTD();
     virtual ~HsmEventDispatcherSTD();
 
-    virtual HandlerID_t registerEventHandler(const EventHandlerFunc_t& handler) override;
-    virtual void unregisterEventHandler(const HandlerID_t handlerID) override;
-    virtual void emitEvent() override;
+    virtual void emitEvent(const HandlerID_t handlerID) override;
 
     virtual bool start() override;
     void stop();
@@ -34,18 +28,13 @@ public:
     void join();
 
 protected:
-    void unregisterAllEventHandlers();
     void doDispatching();
 
 private:
     std::thread mDispatcherThread;
-    std::mutex mEmitSync;
     std::condition_variable mEmitEvent;
     std::mutex mStartupSync;
     std::condition_variable mStartupEvent;
-    std::mutex mHandlersSync;
-    std::map<HandlerID_t, EventHandlerFunc_t> mEventHandlers;
-    int mPendingEmitCount = 0;
     bool mStopDispatcher = false;
 };