+++ src/HsmEventDispatcherGLibmm.cpp
@@ -3,6 +3,7 @@
 
 #include "hsmcpp/HsmEventDispatcherGLibmm.hpp"
 #include "hsmcpp/logging.hpp"
+#include <algorithm>
 
 namespace hsmcpp
 {
@@ -28,62 +29,37 @@ HsmEventDispatcherGLibmm::HsmEventDispatcherGLibmm(const Glib::RefPtr<Glib::Main
 HsmEventDispatcherGLibmm::~HsmEventDispatcherGLibmm()
 {
     __HSM_TRACE_CALL_DEBUG__();
+
+    if (true == mDispatcherConnection.connected())
+    {
+        mDispatcherConnection.disconnect();
+    }
     
     unregisterAllTimerHandlers();
     unregisterAllEventHandlers();
 }
 
-HandlerID_t HsmEventDispatcherGLibmm::registerEventHandler(const EventHandlerFunc_t& handler)
-{
-    __HSM_TRACE_CALL_DEBUG__();
-    HandlerID_t id = INVALID_HSM_DISPATCHER_HANDLER_ID;
-
-    if (mDispatcher)
-    {
-        id = getNextHandlerID();
-        mEventHandlers.emplace(id, mDispatcher->connect(handler));
-    }
-
-    return id;
-}
-
-void HsmEventDispatcherGLibmm::unregisterEventHandler(const HandlerID_t handlerID)
-{
-    __HSM_TRACE_CALL_DEBUG_ARGS__("handlerId=%d", handlerID);
-    auto it = mEventHandlers.find(handlerID);
-
-    if (it != mEventHandlers.end())
-    {
-        it->second.disconnect();
-        mEventHandlers.erase(it);
-    }
-}
-
-void HsmEventDispatcherGLibmm::emitEvent()
+void HsmEventDispatcherGLibmm::emitEvent(const HandlerID_t handlerID)
 {
     __HSM_TRACE_CALL_DEBUG__();
     if (mDispatcher)
     {
+        HsmEventDispatcherBase::emitEvent(handlerID);
         mDispatcher->emit();
     }
 }
 
 bool HsmEventDispatcherGLibmm::start()
 {
-    // do nothing
-    // NOTE: in case of GLib based dispatcher implementation it's expected
-    //       that application will run GLib MainLoop in it's own code.
-    return true;
-}
+    bool result = false;
 
-void HsmEventDispatcherGLibmm::unregisterAllEventHandlers()
-{
-    for (auto it = mEventHandlers.begin(); it != mEventHandlers.end(); ++it)
+    if (mDispatcher)
     {
-        it->second.disconnect();
+        mDispatcherConnection = mDispatcher->connect(sigc::mem_fun(this, &HsmEventDispatcherGLibmm::onDispatchEvents));
+        result = true;
     }
 
-    mEventHandlers.clear();
+    return result;
 }
 
 void HsmEventDispatcherGLibmm::unregisterAllTimerHandlers()
@@ -126,6 +102,13 @@ void HsmEventDispatcherGLibmm::stopTimerImpl(const TimerID_t timerID)
         it->second.disconnect();
         mTimerHandlers.erase(it);
     }
+}
+
+void HsmEventDispatcherGLibmm::onDispatchEvents()
+{
+    __HSM_TRACE_CALL_DEBUG__();
+
+    HsmEventDispatcherBase::dispatchPendingEvents();
 }
 
 bool HsmEventDispatcherGLibmm::onTimerEvent(const TimerID_t timerID)